+++ core/model/event.go
@@ -29,6 +29,24 @@ type Event struct {
 	ToMembersList []ToMember `json:"to_members" bson:"to_members"` // nil or empty means everyone; non-empty means visible to those user ids and admins
 } // @name Event
 
+// HasToMembersList Checks if the ToMembersList is not empty
+func (e Event) HasToMembersList() bool {
+	return len(e.ToMembersList) > 0
+}
+
+// HasToMemberUser Checks if user with identifier exists whithin the ToMembers ACL
+func (e Event) HasToMemberUser(userID *string, externalID *string) bool {
+	for _, toMember := range e.ToMembersList {
+		if userID != nil && toMember.UserID == *userID {
+			return true
+		}
+		if externalID != nil && toMember.ExternalID == *externalID {
+			return true
+		}
+	}
+	return false
+}
+
 // GetMembersAsUserIDs Gets all members as list of user ids
 func (e Event) GetMembersAsUserIDs(skipUserID *string) []string {
 	var userIDs []string
@@ -57,3 +75,13 @@ func (e Event) GetMembersAsNotificationRecipients(skipUserID *string) []notifica
 	}
 	return recipients
 }
+
+// GroupEventFilter event filter wrapper
+type GroupEventFilter struct {
+	Limit  *int64 `json:"limit,omitempty"`
+	Offset *int64 `json:"offset,omitempty"`
+
+	StartTimeAfter *int64 `json:"start_time_after,omitempty"`
+
+	StartTimeBefore *int64 `json:"start_time_before,omitempty"`
+} // @name GroupEventFilter