+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Integrate Calendar BB for group events [#392](https://github.com/rokwire/groups-building-block/issues/392)
+- Create an adaptor for requesting Calendar BB for dealing with group events [#391](https://github.com/rokwire/groups-building-block/issues/391)
+
 ## [1.21.0] - 2023-09-19
 ### Changed
 - Updated libraries and docker container due to vulnerabilities along with the original ticket[#386](https://github.com/rokwire/groups-building-block/issues/386)